improved validation and returned exception errors as json

+++ database/migrations/2016_09_03_180035_create_product_table.php
@@ -17,7 +17,7 @@ class CreateProductTable extends Migration
             $table->increments('id');
             $table->string('name', 100);
             $table->string('summary', 500);
-            $table->integer('price');
+            $table->decimal('price');
             $table->integer('stock');
             $table->timestamps();
         });
